pyproject.toml
[project]
name = "imap-icloud-migration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioimaplib>=1.0.0",
    "google-api-python-client>=2.120",
    "google-auth>=2.28",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.2",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "rich>=14.2.0",
    "typer>=0.12",
]

[project.scripts]
imap-icloud-migration = "imap_icloud_migration.__main__:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 100
extend-exclude = [".venv", "data", ".secrets", ".history"]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "UP", # pyupgrade
]
ignore = [
  "B008", # Typer uses Option(...) in defaults by design
]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = "(^\\.venv/|^data/|^\\.secrets/)"

[[tool.mypy.overrides]]
module = [
  "aioimaplib",
  "googleapiclient.*",
  "google_auth_oauthlib.*",
  "google.auth.*",
  "google.oauth2.*",
]
ignore_missing_imports = true
disable_error_code = ["import-untyped", "no-untyped-call"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
]
